+++ test/native-contracts/Program.cs
@@ -7,12 +7,12 @@ using SimpleJSON;
 
 namespace Neo
 {
-    class UInt160 {}
-    class UInt256 {}
+    class UInt160 { }
+    class UInt256 { }
 
     namespace Cryptography.ECC
     {
-        class ECPoint {}
+        class ECPoint { }
     }
 }
 
@@ -33,7 +33,7 @@ namespace native_contracts
             var response = await client.PostAsync("http://seed5t.neo.org:20332", requestContent);
             var responseBytes = await response.Content.ReadAsByteArrayAsync();
             var responseJson = SimpleJSON.JSONObject.Parse(Encoding.UTF8.GetString(responseBytes));
-            
+
             if (!System.IO.Directory.Exists("./out")) System.IO.Directory.CreateDirectory("./out");
             var results = (JSONArray)responseJson["result"];
             foreach (var result in results)